Revert "Changes to sell module"
This reverts commit 7a33f269

+++ ui/src/main/java/kg/nurtelecom/ofd/sell/CustomCardView.kt
@@ -8,7 +8,6 @@ import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import com.google.android.material.card.MaterialCardView
 import kg.nurtelecom.ui.databinding.ViewCardBinding
-import java.math.BigDecimal
 
 class CustomCardView(context: Context, attrs: AttributeSet? = null) :
     MaterialCardView(context, attrs) {
@@ -20,7 +19,7 @@ class CustomCardView(context: Context, attrs: AttributeSet? = null) :
         binding.cardTitle.setText(title)
     }
 
-    fun setCardContent(value: BigDecimal) {
+    fun setCardContent(value: Double) {
         binding.cardContent.text = value.toString()
     }
